+++ boards/stm32l432kc/src/bin/levitation.rs
@@ -9,7 +9,8 @@ use embassy_stm32::timer::low_level::CountingMode;
 use embassy_stm32::timer::simple_pwm::{PwmPin, SimplePwm};
 use embassy_stm32::timer::Channel;
 use embassy_time::Instant;
-use hyped_control::pid_controller::{Pi, PiController, PiGain, Pid, PidController, PidGain};
+use hyped_control::pid_controller::{Pid, PidController, PidGain};
+use hyped_control::pi_controller::{Pi, PiController, PiGain};
 
 use {defmt_rtt as _, panic_probe as _};
